+++ crates/configuration/src/configuration.rs
@@ -2,6 +2,7 @@ use std::{collections::BTreeMap, path::Path};
 
 use anyhow::{anyhow, ensure};
 use itertools::Itertools;
+use mongodb_support::ExtendedJsonMode;
 use ndc_models as ndc;
 use serde::{Deserialize, Serialize};
 
@@ -189,11 +190,16 @@ impl Configuration {
     }
 }
 
-#[derive(Copy, Clone, Debug, Default, Deserialize, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ConfigurationOptions {
-    // Options for introspection
+    /// Options for introspection
     pub introspection_options: ConfigurationIntrospectionOptions,
+
+    /// Options that affect how BSON data from MongoDB is translated to JSON in GraphQL query
+    /// responses.
+    #[serde(default)]
+    pub serialization_options: ConfigurationSerializationOptions,
 }
 
 #[derive(Copy, Clone, Debug, Deserialize, Serialize)]
@@ -217,6 +223,15 @@ impl Default for ConfigurationIntrospectionOptions {
             all_schema_nullable: true,
         }
     }
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ConfigurationSerializationOptions {
+    /// Extended JSON has two modes: canonical and relaxed. This option determines which mode is
+    /// used for output. This setting has no effect on inputs (query arguments, etc.).
+    #[serde(default)]
+    pub extended_json_mode: ExtendedJsonMode,
 }
 
 fn merge_object_types<'a>(