+++ crates/mongodb-agent-common/src/mongo_query_plan/mod.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeMap;
 use configuration::{
     native_mutation::NativeMutation, native_query::NativeQuery, Configuration, MongoScalarType,
 };
-use mongodb_support::EXTENDED_JSON_TYPE_NAME;
+use mongodb_support::{ExtendedJsonMode, EXTENDED_JSON_TYPE_NAME};
 use ndc_models as ndc;
 use ndc_query_plan::{ConnectorTypes, QueryContext, QueryPlanError};
 
@@ -17,6 +17,10 @@ pub use ndc_query_plan::OrderByTarget;
 pub struct MongoConfiguration(pub Configuration);
 
 impl MongoConfiguration {
+    pub fn extended_json_mode(&self) -> ExtendedJsonMode {
+        self.0.options.serialization_options.extended_json_mode
+    }
+
     pub fn native_queries(&self) -> &BTreeMap<String, NativeQuery> {
         &self.0.native_queries
     }